backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.11",
    "fastapi>=0.112.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.1",
    "openai>=1.40.0",
]
